Change access modifiers of EventProcessorClient fields
Currently, a developer will find EventProcessorClient hard to extend: you can either use EventProcessorClient as-is, or directly extend EventProcessor<TPartition> and manage a lot of unnecessary code and dependencies.
An example scenario: I wish to override OnProcessingEventBatchAsync so instead of calling _processEventAsync sequentially for each event, it will handle the events in parallel. This is a common scenario for handling messages.
Currently we cannot do so, because there is no access for methods like UpdateCheckpointAsync.